♻️ 收敛目录同步与文件管理器实现
