+++ frontend/src/components/terminal/FileManagerPanel.tsx
@@ -1,286 +1,20 @@
-import { useState, useEffect, useCallback, useRef, useMemo, useLayoutEffect } from "react";
-import { createPortal } from "react-dom";
+import { useCallback, useEffect, useMemo, useRef, useState, type CSSProperties } from "react";
 import { useTranslation } from "react-i18next";
-import {
-  Folder,
-  File,
-  Loader2,
-  ArrowUp,
-  Home,
-  RefreshCw,
-  Upload,
-  Download,
-  FolderDown,
-  Trash2,
-  X,
-  CheckCircle2,
-  XCircle,
-  FolderOpen,
-} from "lucide-react";
-import { cn, Button, Input, ScrollArea, ConfirmDialog } from "@opskat/ui";
-import { ChangeSSHDirectory, SFTPListDir, SFTPGetwd, SFTPDelete } from "../../../wailsjs/go/app/App";
-import { OnFileDrop, OnFileDropOff } from "../../../wailsjs/runtime/runtime";
-import { useSFTPStore } from "@/stores/sftpStore";
-import { useTerminalStore } from "@/stores/terminalStore";
+import { Upload } from "lucide-react";
+import { cn, ConfirmDialog } from "@opskat/ui";
+import { SFTPDelete, SFTPGetwd } from "../../../wailsjs/go/app/App";
 import { sftp_svc } from "../../../wailsjs/go/models";
-import { toast } from "sonner";
-
-// --- Helpers ---
-
-function formatBytes(bytes: number): string {
-  if (bytes === 0) return "0 B";
-  const units = ["B", "KB", "MB", "GB", "TB"];
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  return (bytes / Math.pow(1024, i)).toFixed(i > 0 ? 1 : 0) + " " + units[i];
-}
-
-function formatDate(timestamp: number): string {
-  const d = new Date(timestamp * 1000);
-  return d.toLocaleDateString(undefined, {
-    month: "2-digit",
-    day: "2-digit",
-  });
-}
-
-function normalizeRemotePath(basePath: string, nextPath: string): string {
-  const raw = nextPath.trim();
-  if (!raw) return basePath || "/";
-  const combined = raw.startsWith("/") ? raw : `${basePath === "/" ? "" : basePath}/${raw}`;
-  const parts = combined.split("/");
-  const normalized: string[] = [];
-  for (const part of parts) {
-    if (!part || part === ".") continue;
-    if (part === "..") {
-      normalized.pop();
-      continue;
-    }
-    normalized.push(part);
-  }
-  return "/" + normalized.join("/");
-}
-
-const HANDLE_PX = 4;
-
-// --- Context Menu ---
-
-interface CtxMenuState {
-  x: number;
-  y: number;
-  entry: sftp_svc.FileEntry | null;
-}
-
-function FloatingMenu({
-  ctx,
-  onAction,
-  onClose,
-}: {
-  ctx: CtxMenuState;
-  onAction: (action: string) => void;
-  onClose: () => void;
-}) {
-  const { t } = useTranslation();
-  const ref = useRef<HTMLDivElement>(null);
-  const [pos, setPos] = useState({ top: 0, left: 0 });
-  const [visible, setVisible] = useState(false);
-  const [interactive, setInteractive] = useState(false);
-
-  useLayoutEffect(() => {
-    if (!ref.current) return;
-    const rect = ref.current.getBoundingClientRect();
-    const vw = window.innerWidth;
-    const vh = window.innerHeight;
-    let left = ctx.x + 2;
-    let top = ctx.y + 2;
-    if (left + rect.width > vw) left = ctx.x - rect.width - 2;
-    if (top + rect.height > vh) top = ctx.y - rect.height - 2;
-    left = Math.max(4, Math.min(left, vw - rect.width - 4));
-    top = Math.max(4, Math.min(top, vh - rect.height - 4));
-    setPos({ top, left });
-    setVisible(true);
-  }, [ctx.x, ctx.y]);
-
-  useEffect(() => {
-    const timer = setTimeout(() => setInteractive(true), 150);
-    return () => clearTimeout(timer);
-  }, []);
-
-  useEffect(() => {
-    const handleKey = (e: KeyboardEvent) => {
-      if (e.key === "Escape") onClose();
-    };
-    const handlePointer = (e: PointerEvent) => {
-      if (ref.current?.contains(e.target as Node)) return;
-      onClose();
-    };
-    const timer = setTimeout(() => {
-      document.addEventListener("pointerdown", handlePointer, true);
-    }, 50);
-    document.addEventListener("keydown", handleKey);
-    return () => {
-      clearTimeout(timer);
-      document.removeEventListener("pointerdown", handlePointer, true);
-      document.removeEventListener("keydown", handleKey);
-    };
-  }, [onClose]);
-
-  const item = (action: string, icon: React.ReactNode, label: string, variant?: "destructive") => (
-    <div
-      key={action}
-      className={cn(
-        "flex items-center gap-2 rounded-sm px-2 py-1.5 text-sm cursor-default select-none [&_svg]:shrink-0 [&_svg:not([class*='size-'])]:size-4",
-        variant === "destructive"
-          ? "text-destructive hover:bg-destructive/10 [&_svg]:text-destructive"
-          : "hover:bg-accent hover:text-accent-foreground [&_svg:not([class*='text-'])]:text-muted-foreground"
-      )}
-      onClick={() => onAction(action)}
-    >
-      {icon}
-      {label}
-    </div>
-  );
-
-  return createPortal(
-    <div
-      ref={ref}
-      className={cn(
-        "z-50 min-w-[10rem] overflow-hidden rounded-md border bg-popover p-1 text-popover-foreground shadow-md",
-        visible && "animate-in fade-in-0 zoom-in-95"
-      )}
-      style={{
-        position: "fixed",
-        top: pos.top,
-        left: pos.left,
-        visibility: visible ? "visible" : "hidden",
-        pointerEvents: interactive ? "auto" : "none",
-      }}
-    >
-      {ctx.entry ? (
-        ctx.entry.isDir ? (
-          <>
-            {item("open", <FolderOpen />, t("sftp.openFolder"))}
-            {item("downloadDir", <FolderDown />, t("sftp.downloadDir"))}
-            <div className="-mx-1 my-1 h-px bg-border" />
-            {item("delete", <Trash2 />, t("action.delete"), "destructive")}
-          </>
-        ) : (
-          <>
-            {item("download", <Download />, t("sftp.download"))}
-            <div className="-mx-1 my-1 h-px bg-border" />
-            {item("delete", <Trash2 />, t("action.delete"), "destructive")}
-          </>
-        )
-      ) : (
-        <>
-          {item("upload", <Upload />, t("sftp.upload"))}
-          {item("uploadDir", <Upload />, t("sftp.uploadDir"))}
-          <div className="-mx-1 my-1 h-px bg-border" />
-          {item("refresh", <RefreshCw />, t("sftp.refresh"))}
-        </>
-      )}
-    </div>,
-    document.body
-  );
-}
-
-// --- Transfer Row ---
-
-function TransferRow({
-  transfer,
-}: {
-  transfer: {
-    transferId: string;
-    direction: "upload" | "download";
-    currentFile: string;
-    bytesDone: number;
-    bytesTotal: number;
-    speed: number;
-    status: "active" | "done" | "error" | "cancelled";
-    error?: string;
-  };
-}) {
-  const cancelTransfer = useSFTPStore((s) => s.cancelTransfer);
-  const clearTransfer = useSFTPStore((s) => s.clearTransfer);
-
-  const percent = transfer.bytesTotal > 0 ? Math.round((transfer.bytesDone / transfer.bytesTotal) * 100) : 0;
-
-  const fileName = transfer.currentFile ? transfer.currentFile.split("/").pop() || transfer.currentFile : "";
-
-  return (
-    <div className="flex items-center gap-1.5 text-[11px]">
-      <div className="shrink-0">
-        {transfer.status === "active" && <Loader2 className="h-3 w-3 animate-spin text-primary" />}
-        {transfer.status === "done" && <CheckCircle2 className="h-3 w-3 text-green-500" />}
-        {(transfer.status === "error" || transfer.status === "cancelled") && (
-          <XCircle className="h-3 w-3 text-destructive" />
-        )}
-      </div>
-
-      <div className="flex-1 min-w-0">
-        <div className="flex items-center gap-1">
-          {transfer.direction === "upload" ? (
-            <Upload className="h-2.5 w-2.5 text-muted-foreground shrink-0" />
-          ) : (
-            <Download className="h-2.5 w-2.5 text-muted-foreground shrink-0" />
-          )}
-          <span className="truncate">{fileName}</span>
-          {transfer.status === "active" && <span className="shrink-0 text-muted-foreground ml-auto">{percent}%</span>}
-        </div>
-        {transfer.status === "active" && (
-          <div className="h-1 rounded-full bg-muted overflow-hidden mt-0.5">
-            <div
-              className="h-full rounded-full bg-primary transition-all duration-300"
-              style={{ width: `${percent}%` }}
-            />
-          </div>
-        )}
-        {transfer.status === "error" && transfer.error && (
-          <span className="text-destructive truncate block text-[10px]" title={transfer.error}>
-            {transfer.error}
-          </span>
-        )}
-      </div>
-
-      <Button
-        variant="ghost"
-        size="icon-xs"
-        className="shrink-0 h-4 w-4"
-        onClick={() =>
-          transfer.status === "active" ? cancelTransfer(transfer.transferId) : clearTransfer(transfer.transferId)
-        }
-      >
-        <X className="h-2.5 w-2.5" />
-      </Button>
-    </div>
-  );
-}
-
-function formatSyncError(t: (key: string) => string, error: unknown): string {
-  const message = String(error);
-  switch (message) {
-    case "DIRSYNC_UNSUPPORTED":
-      return t("sftp.sync.unsupported");
-    case "DIRSYNC_CWD_UNKNOWN":
-      return t("sftp.sync.cwdUnknown");
-    case "DIRSYNC_BUSY":
-      return t("sftp.sync.busy");
-    case "DIRSYNC_PENDING":
-      return t("sftp.sync.pending");
-    case "DIRSYNC_TIMEOUT":
-      return t("sftp.sync.timeout");
-    case "DIRSYNC_NOT_FOUND":
-      return t("sftp.sync.notFound");
-    case "DIRSYNC_NOT_DIRECTORY":
-      return t("sftp.sync.notDirectory");
-    case "DIRSYNC_ACCESS_DENIED":
-      return t("sftp.sync.accessDenied");
-    case "DIRSYNC_REMOTE_CHDIR_FAILED":
-      return t("sftp.sync.remoteFailed");
-    default:
-      return message;
-  }
-}
-
-// --- Main Panel ---
+import { useSFTPStore } from "@/stores/sftpStore";
+import { FileList } from "./file-manager/FileList";
+import { FloatingMenu } from "./file-manager/FloatingMenu";
+import { PathToolbar } from "./file-manager/PathToolbar";
+import { TransferSection } from "./file-manager/TransferSection";
+import { type DeleteTarget, type CtxMenuState } from "./file-manager/types";
+import { useFileManagerDirectory } from "./file-manager/useFileManagerDirectory";
+import { useNativeFileDrop } from "./file-manager/useNativeFileDrop";
+import { useResizeHandle } from "./file-manager/useResizeHandle";
+import { useTerminalDirectorySync } from "./file-manager/useTerminalDirectorySync";
+import { getEntryPath, getParentPath, HANDLE_PX } from "./file-manager/utils";
 
 interface FileManagerPanelProps {
   tabId: string;
@@ -292,141 +26,63 @@ interface FileManagerPanelProps {
 
 export function FileManagerPanel({ tabId, sessionId, isOpen, width, onWidthChange }: FileManagerPanelProps) {
   const { t } = useTranslation();
+  const {
+    currentPath,
+    currentPathRef,
+    entries,
+    error,
+    loading,
+    loadDir,
+    pathInput,
+    selected,
+    setError,
+    setPathInput,
+    setSelected,
+    storedPath,
+  } = useFileManagerDirectory(tabId, sessionId);
 
-  const storedPath = useSFTPStore((s) => s.fileManagerPaths[tabId]);
-  const currentPath = storedPath || "/";
-  const setCurrentPath = useSFTPStore((s) => s.setFileManagerPath);
-  const [pathInput, setPathInput] = useState(currentPath);
-  const [entries, setEntries] = useState<sftp_svc.FileEntry[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [selected, setSelected] = useState<string | null>(null);
+  const {
+    directoryFollowMode,
+    navigateToPath,
+    paneConnected,
+    sessionSync,
+    syncPanelFromTerminal,
+    syncTerminalToPath,
+    toggleFollowMode,
+  } = useTerminalDirectorySync({
+    currentPathRef,
+    loadDir,
+    sessionId,
+    tabId,
+  });
 
-  // Context menu
   const [ctxMenu, setCtxMenu] = useState<CtxMenuState | null>(null);
-
-  // Delete confirmation
-  const [deleteTarget, setDeleteTarget] = useState<{
-    path: string;
-    name: string;
-    isDir: boolean;
-  } | null>(null);
-
-  // Drag overlay
-  const [isDragOver, setIsDragOver] = useState(false);
-
-  // Resize
-  const [isResizing, setIsResizing] = useState(false);
-
-  // Track whether initial load happened
+  const [deleteTarget, setDeleteTarget] = useState<DeleteTarget | null>(null);
   const loadedRef = useRef(false);
   const lastSessionRef = useRef<string | null>(null);
-  const loadRequestRef = useRef(0);
-
-  // Refs
   const panelRef = useRef<HTMLDivElement>(null);
-  const currentPathRef = useRef(currentPath);
-  currentPathRef.current = currentPath;
 
-  // Store
   const startUpload = useSFTPStore((s) => s.startUpload);
   const startUploadDir = useSFTPStore((s) => s.startUploadDir);
   const startUploadFile = useSFTPStore((s) => s.startUploadFile);
   const startDownload = useSFTPStore((s) => s.startDownload);
   const startDownloadDir = useSFTPStore((s) => s.startDownloadDir);
   const allTransfers = useSFTPStore((s) => s.transfers);
-  const clearCompletedForSession = useSFTPStore((s) => s.clearCompletedForSession);
-  const directoryFollowMode = useTerminalStore((s) => s.tabData[tabId]?.directoryFollowMode ?? "off");
-  const setDirectoryFollowMode = useTerminalStore((s) => s.setDirectoryFollowMode);
-  const sessionSync = useTerminalStore((s) => s.sessionSync[sessionId]);
-  const paneConnected = useTerminalStore((s) => s.tabData[tabId]?.panes[sessionId]?.connected ?? false);
 
   const sessionTransfers = useMemo(
-    () => Object.values(allTransfers).filter((t) => t.sessionId === sessionId),
+    () => Object.values(allTransfers).filter((transfer) => transfer.sessionId === sessionId),
     [allTransfers, sessionId]
   );
 
-  // === File browsing ===
-  const loadDir = useCallback(
-    async (dirPath: string) => {
-      const requestId = ++loadRequestRef.current;
-      const normalizedPath = normalizeRemotePath(currentPathRef.current, dirPath);
-      setLoading(true);
-      setError(null);
-      setSelected(null);
-      try {
-        const result = await SFTPListDir(sessionId, normalizedPath);
-        if (requestId !== loadRequestRef.current) return false;
-        setEntries(result || []);
-        setCurrentPath(tabId, normalizedPath);
-        setPathInput(normalizedPath);
-        return true;
-      } catch (e) {
-        if (requestId !== loadRequestRef.current) return false;
-        setError(String(e));
-        return false;
-      } finally {
-        if (requestId === loadRequestRef.current) {
-          setLoading(false);
-        }
-      }
-    },
-    [sessionId, setCurrentPath, tabId]
-  );
+  const isDragOver = useNativeFileDrop({
+    currentPathRef,
+    isOpen,
+    panelRef,
+    sessionId,
+    startUploadFile,
+  });
+  const { handleResizeStart, isResizing, outerRef } = useResizeHandle({ onWidthChange, panelRef, width });
 
-  useEffect(() => {
-    setPathInput(currentPath);
-  }, [currentPath]);
-
-  const showSyncError = useCallback(
-    (err: unknown) => {
-      toast.error(formatSyncError(t, err));
-    },
-    [t]
-  );
-
-  const syncPanelFromTerminal = useCallback(async () => {
-    if (!sessionSync) {
-      showSyncError("DIRSYNC_CWD_UNKNOWN");
-      return false;
-    }
-    if (!sessionSync.supported) {
-      showSyncError("DIRSYNC_UNSUPPORTED");
-      return false;
-    }
-    if (!sessionSync.cwdKnown || !sessionSync.cwd) {
-      showSyncError("DIRSYNC_CWD_UNKNOWN");
-      return false;
-    }
-    return loadDir(sessionSync.cwd);
-  }, [loadDir, sessionSync, showSyncError]);
-
-  const syncTerminalToPath = useCallback(
-    async (targetPath: string) => {
-      try {
-        await ChangeSSHDirectory(sessionId, targetPath);
-        return true;
-      } catch (e) {
-        showSyncError(e);
-        return false;
-      }
-    },
-    [sessionId, showSyncError]
-  );
-
-  const navigateToPath = useCallback(
-    async (dirPath: string) => {
-      const targetPath = normalizeRemotePath(currentPathRef.current, dirPath);
-      if (directoryFollowMode === "always") {
-        const changed = await syncTerminalToPath(targetPath);
-        if (!changed) return false;
-      }
-      return loadDir(targetPath);
-    },
-    [directoryFollowMode, loadDir, syncTerminalToPath]
-  );
-
-  // Load directory only once on first open, or when session changes
   useEffect(() => {
     if (!sessionId) return;
     if (lastSessionRef.current !== sessionId) {
@@ -457,8 +113,9 @@ export function FileManagerPanel({ tabId, sessionId, isOpen, width, onWidthChang
     void loadDir(sessionSync.cwd);
   }, [currentPath, directoryFollowMode, isOpen, loadDir, sessionSync?.cwd, sessionSync?.cwdKnown]);
 
-  // Auto-refresh after upload completes
-  const doneUploadCount = sessionTransfers.filter((t) => t.status === "done" && t.direction === "upload").length;
+  const doneUploadCount = sessionTransfers.filter((transfer) => {
+    return transfer.status === "done" && transfer.direction === "upload";
+  }).length;
   const prevDoneCount = useRef(0);
   useEffect(() => {
     if (doneUploadCount > prevDoneCount.current) {
@@ -468,118 +125,19 @@ export function FileManagerPanel({ tabId, sessionId, isOpen, width, onWidthChang
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [doneUploadCount]);
 
-  // === Navigation helpers ===
-  const getFullPath = useCallback(
-    (entry: sftp_svc.FileEntry) => (currentPath === "/" ? "/" + entry.name : currentPath + "/" + entry.name),
-    [currentPath]
-  );
+  const getFullPath = useCallback((entry: sftp_svc.FileEntry) => getEntryPath(currentPath, entry), [currentPath]);
 
-  const goUp = () => {
+  const goUp = useCallback(() => {
     if (currentPath === "/") return;
-    const parent = currentPath.replace(/\/[^/]+\/?$/, "") || "/";
-    void navigateToPath(parent);
-  };
+    void navigateToPath(getParentPath(currentPath));
+  }, [currentPath, navigateToPath]);
 
-  const goHome = () => {
+  const goHome = useCallback(() => {
     SFTPGetwd(sessionId)
       .then((home) => navigateToPath(home || "/"))
       .catch(() => navigateToPath("/"));
-  };
+  }, [navigateToPath, sessionId]);
 
-  const toggleFollowMode = useCallback(async () => {
-    if (directoryFollowMode === "always") {
-      setDirectoryFollowMode(tabId, "off");
-      return;
-    }
-
-    if (sessionSync?.busy) {
-      showSyncError("DIRSYNC_BUSY");
-      return;
-    }
-
-    const synced = await syncPanelFromTerminal();
-    if (!synced) return;
-    setDirectoryFollowMode(tabId, "always");
-  }, [directoryFollowMode, sessionSync?.busy, setDirectoryFollowMode, showSyncError, syncPanelFromTerminal, tabId]);
-
-  // === Drag and drop (Wails native) — only when panel is open ===
-  useEffect(() => {
-    if (!isOpen) return;
-    const handler = (_x: number, _y: number, paths: string[]) => {
-      setIsDragOver(false);
-      for (const path of paths) {
-        startUploadFile(sessionId, path, currentPathRef.current + "/");
-      }
-    };
-    OnFileDrop(handler, true);
-    return () => {
-      OnFileDropOff();
-    };
-  }, [isOpen, sessionId, startUploadFile]);
-
-  // Track Wails drop-target class via MutationObserver for drag overlay
-  useEffect(() => {
-    const el = panelRef.current;
-    if (!el || !isOpen) return;
-    const observer = new MutationObserver(() => {
-      setIsDragOver(el.classList.contains("wails-drop-target-active"));
-    });
-    observer.observe(el, { attributes: true, attributeFilter: ["class"] });
-    return () => observer.disconnect();
-  }, [isOpen]);
-
-  // === Resize handle (with body cursor lock) ===
-  // 拖拽期间通过 ref 直接改 DOM 宽度（rAF 合批），松手才 setState —— 避免把 MainPanel 子树 60Hz 重渲
-  const outerRef = useRef<HTMLDivElement>(null);
-  const handleResizeStart = useCallback(
-    (e: React.MouseEvent) => {
-      e.preventDefault();
-      setIsResizing(true);
-      const startX = e.clientX;
-      const startWidth = width;
-      const prevCursor = document.body.style.cursor;
-      document.body.style.cursor = "col-resize";
-
-      const outer = outerRef.current;
-      const prevOuterTransition = outer?.style.transition ?? "";
-      // 禁 width transition，否则拖拽时宽度有延迟
-      if (outer) outer.style.transition = "none";
-
-      let pending = startWidth;
-      let rafId: number | null = null;
-      const flushToDom = () => {
-        rafId = null;
-        if (panelRef.current) panelRef.current.style.width = `${pending}px`;
-        if (outer) outer.style.width = `${pending + HANDLE_PX}px`;
-      };
-
-      const onMove = (e: MouseEvent) => {
-        pending = startWidth + (startX - e.clientX);
-        if (rafId == null) rafId = requestAnimationFrame(flushToDom);
-      };
-      const onUp = () => {
-        if (rafId != null) {
-          cancelAnimationFrame(rafId);
-          rafId = null;
-        }
-        setIsResizing(false);
-        document.body.style.cursor = prevCursor;
-        document.removeEventListener("mousemove", onMove);
-        document.removeEventListener("mouseup", onUp);
-        if (panelRef.current) panelRef.current.style.width = "";
-        if (outer) {
-          outer.style.width = "";
-          outer.style.transition = prevOuterTransition;
-        }
-        onWidthChange(pending);
-      };
-      document.addEventListener("mousemove", onMove);
-      document.addEventListener("mouseup", onUp);
-    },
-    [width, onWidthChange]
-  );
-
-  // === Delete ===
   const handleDelete = useCallback(async () => {
     if (!deleteTarget) return;
     try {
@@ -590,9 +148,8 @@ export function FileManagerPanel({ tabId, sessionId, isOpen, width, onWidthChang
     } finally {
       setDeleteTarget(null);
     }
-  }, [deleteTarget, loadDir, sessionId]);
+  }, [currentPathRef, deleteTarget, loadDir, sessionId, setError]);
 
-  // === Context menu actions ===
   const handleCtxAction = useCallback(
     (action: string) => {
       if (!ctxMenu) return;
@@ -631,33 +188,23 @@ export function FileManagerPanel({ tabId, sessionId, isOpen, width, onWidthChang
     },
     [
       ctxMenu,
-      sessionId,
-      loadDir,
+      currentPath,
+      currentPathRef,
       getFullPath,
+      loadDir,
       navigateToPath,
-      startUpload,
-      startUploadDir,
+      sessionId,
       startDownload,
       startDownloadDir,
+      startUpload,
+      startUploadDir,
     ]
-  );
-
-  // Sort entries: directories first, then alphabetical
-  const sortedEntries = useMemo(
-    () =>
-      [...entries].sort((a, b) => {
-        if (a.isDir && !b.isDir) return -1;
-        if (!a.isDir && b.isDir) return 1;
-        return a.name.localeCompare(b.name);
-      }),
-    [entries]
   );
 
   const totalWidth = width + HANDLE_PX;
 
   return (
     <>
-      {/* Animated outer wrapper */}
       <div
         ref={outerRef}
         className="shrink-0 overflow-hidden transition-[width] duration-200 ease-out"
@@ -667,7 +214,6 @@ export function FileManagerPanel({ tabId, sessionId, isOpen, width, onWidthChang
         }}
       >
         <div className="flex h-full" style={{ minWidth: totalWidth }}>
-          {/* Resize handle */}
           <div
             className={cn(
               "w-1 cursor-col-resize hover:bg-primary/20 transition-colors shrink-0",
@@ -676,7 +222,6 @@ export function FileManagerPanel({ tabId, sessionId, isOpen, width, onWidthChang
             onMouseDown={handleResizeStart}
           />
 
-          {/* Panel */}
           <div
             ref={panelRef}
             className="flex flex-col border-l bg-background relative overflow-hidden"
@@ -684,10 +229,9 @@ export function FileManagerPanel({ tabId, sessionId, isOpen, width, onWidthChang
               {
                 width,
                 "--wails-drop-target": isOpen ? "drop" : undefined,
-              } as React.CSSProperties
+              } as CSSProperties
             }
           >
-            {/* Drop overlay */}
             {isDragOver && (
               <div className="pointer-events-none absolute inset-0 z-10 flex items-center justify-center bg-primary/5 border-2 border-dashed border-primary/30 rounded animate-in fade-in-0 duration-150">
                 <div className="flex flex-col items-center gap-1 text-primary/60">
@@ -697,195 +241,41 @@ export function FileManagerPanel({ tabId, sessionId, isOpen, width, onWidthChang
               </div>
             )}
 
-            {/* Path bar */}
-            <div className="flex items-center gap-0.5 px-1 py-1 border-b shrink-0">
-              <Button
-                variant="ghost"
-                size="xs"
-                className="px-1.5 text-[10px] font-mono"
-                onClick={() => void syncPanelFromTerminal()}
-                title={t("sftp.sync.panelFromTerminal")}
-                aria-label={t("sftp.sync.panelFromTerminal")}
-                disabled={!paneConnected}
-              >
-                T→F
-              </Button>
-              <Button
-                variant="ghost"
-                size="xs"
-                className="px-1.5 text-[10px] font-mono"
-                onClick={() => void syncTerminalToPath(currentPath)}
-                title={t("sftp.sync.terminalFromPanel")}
-                aria-label={t("sftp.sync.terminalFromPanel")}
-                disabled={!paneConnected}
-              >
-                F→T
-              </Button>
-              <Button
-                variant={directoryFollowMode === "always" ? "secondary" : "ghost"}
-                size="xs"
-                className="px-1.5 text-[10px] font-medium"
-                onClick={() => void toggleFollowMode()}
-                title={t("sftp.sync.followToggle")}
-                aria-label={t("sftp.sync.followToggle")}
-                disabled={!paneConnected}
-              >
-                {t("sftp.sync.followShort")}
-              </Button>
-              <Button
-                variant="ghost"
-                size="icon-xs"
-                onClick={goUp}
-                disabled={currentPath === "/"}
-                title={t("sftp.parentDir")}
-              >
-                <ArrowUp className="h-3.5 w-3.5" />
-              </Button>
-              <Button variant="ghost" size="icon-xs" onClick={goHome} title={t("sftp.home")}>
-                <Home className="h-3.5 w-3.5" />
-              </Button>
-              <Input
-                className="h-6 text-xs flex-1 min-w-0"
-                value={pathInput}
-                onChange={(e) => setPathInput(e.target.value)}
-                onKeyDown={(e) => {
-                  if (e.key === "Enter") void navigateToPath(pathInput.trim());
-                }}
-              />
-              <Button
-                variant="ghost"
-                size="icon-xs"
-                onClick={() => void loadDir(currentPathRef.current)}
-                title={t("sftp.refresh")}
-              >
-                <RefreshCw className="h-3.5 w-3.5" />
-              </Button>
-            </div>
+            <PathToolbar
+              currentPath={currentPath}
+              directoryFollowMode={directoryFollowMode}
+              onFollowToggle={() => void toggleFollowMode()}
+              onGoHome={goHome}
+              onGoUp={goUp}
+              onPathInputChange={setPathInput}
+              onPathSubmit={(nextPath) => void navigateToPath(nextPath)}
+              onRefresh={() => void loadDir(currentPathRef.current)}
+              onSyncPanelFromTerminal={() => void syncPanelFromTerminal()}
+              onSyncTerminalToPath={() => void syncTerminalToPath(currentPath)}
+              paneConnected={paneConnected}
+              pathInput={pathInput}
+            />
 
-            {/* File list */}
-            <ScrollArea className="flex-1 min-h-0">
-              <div
-                className="text-xs select-none"
-                onContextMenu={(e) => {
-                  e.preventDefault();
-                  setCtxMenu({ x: e.clientX, y: e.clientY, entry: null });
-                }}
-              >
-                {loading && (
-                  <div className="flex items-center justify-center py-8">
-                    <Loader2 className="h-4 w-4 animate-spin text-muted-foreground" />
-                  </div>
-                )}
-                {error && !loading && (
-                  <div className="flex flex-col items-center justify-center py-8 gap-1 px-2">
-                    <span className="text-destructive text-center text-xs">{t("sftp.loadError")}</span>
-                    <span className="text-muted-foreground text-center break-all text-[10px]">{error}</span>
-                    <Button
-                      variant="outline"
-                      size="xs"
-                      onClick={() => void loadDir(currentPathRef.current)}
-                      className="mt-1"
-                    >
-                      {t("sftp.retry")}
-                    </Button>
-                  </div>
-                )}
-                {!loading && !error && entries.length === 0 && (
-                  <div className="flex items-center justify-center py-8">
-                    <span className="text-muted-foreground">{t("sftp.empty")}</span>
-                  </div>
-                )}
-                {!loading && !error && (
-                  <>
-                    {currentPath !== "/" && (
-                      <div
-                        className="flex items-center gap-1.5 px-2 py-1 cursor-pointer hover:bg-muted/50"
-                        onDoubleClick={goUp}
-                      >
-                        <Folder className="h-3.5 w-3.5 text-muted-foreground shrink-0" />
-                        <span className="flex-1 truncate">..</span>
-                      </div>
-                    )}
-                    {sortedEntries.map((entry) => {
-                      const fullPath = getFullPath(entry);
-                      const isSelected = selected === fullPath;
-                      return (
-                        <div
-                          key={entry.name}
-                          className={cn(
-                            "flex items-center gap-1.5 px-2 py-1 cursor-pointer transition-colors",
-                            isSelected ? "bg-primary/10 text-primary" : "hover:bg-muted/50"
-                          )}
-                          style={{ contentVisibility: "auto", containIntrinsicSize: "auto 28px" }}
-                          onClick={() => setSelected(fullPath)}
-                          onDoubleClick={() => {
-                            if (entry.isDir) void navigateToPath(fullPath);
-                          }}
-                          onContextMenu={(e) => {
-                            e.preventDefault();
-                            e.stopPropagation();
-                            setSelected(fullPath);
-                            setCtxMenu({
-                              x: e.clientX,
-                              y: e.clientY,
-                              entry,
-                            });
-                          }}
-                        >
-                          {entry.isDir ? (
-                            <Folder className="h-3.5 w-3.5 text-primary/70 shrink-0" />
-                          ) : (
-                            <File className="h-3.5 w-3.5 text-muted-foreground shrink-0" />
-                          )}
-                          <span className="flex-1 truncate">{entry.name}</span>
-                          {!entry.isDir && (
-                            <span className="text-muted-foreground shrink-0 text-[10px]">
-                              {formatBytes(entry.size)}
-                            </span>
-                          )}
-                          <span className="text-muted-foreground shrink-0 text-[10px]">
-                            {formatDate(entry.modTime)}
-                          </span>
-                        </div>
-                      );
-                    })}
-                  </>
-                )}
-              </div>
-            </ScrollArea>
+            <FileList
+              currentPath={currentPath}
+              entries={entries}
+              error={error}
+              loading={loading}
+              onGoUp={goUp}
+              onNavigate={(path) => void navigateToPath(path)}
+              onOpenContextMenu={(x, y, entry) => setCtxMenu({ x, y, entry })}
+              onRetry={() => void loadDir(currentPathRef.current)}
+              selected={selected}
+              setSelected={setSelected}
+            />
 
-            {/* Transfer section */}
-            {sessionTransfers.length > 0 && (
-              <div className="border-t shrink-0">
-                <div className="flex items-center justify-between px-2 py-0.5">
-                  <span className="text-[11px] font-medium text-muted-foreground">{t("sftp.transfers")}</span>
-                  <Button
-                    variant="ghost"
-                    size="icon-xs"
-                    className="h-4 w-4"
-                    onClick={() => clearCompletedForSession(sessionId)}
-                    title={t("sftp.clear")}
-                  >
-                    <X className="h-2.5 w-2.5" />
-                  </Button>
-                </div>
-                <ScrollArea className="max-h-28">
-                  <div className="px-2 pb-1 space-y-0.5">
-                    {sessionTransfers.map((transfer) => (
-                      <TransferRow key={transfer.transferId} transfer={transfer} />
-                    ))}
-                  </div>
-                </ScrollArea>
-              </div>
-            )}
+            <TransferSection sessionId={sessionId} transfers={sessionTransfers} />
           </div>
         </div>
       </div>
 
-      {/* Context menu */}
       {ctxMenu && <FloatingMenu ctx={ctxMenu} onAction={handleCtxAction} onClose={() => setCtxMenu(null)} />}
 
-      {/* Delete confirmation */}
       <ConfirmDialog
         open={!!deleteTarget}
         onOpenChange={(open) => {
