+++ frontend/src/components/terminal/file-manager/types.ts
@@ -1,0 +1,13 @@
+import { sftp_svc } from "../../../../wailsjs/go/models";
+
+export interface CtxMenuState {
+  x: number;
+  y: number;
+  entry: sftp_svc.FileEntry | null;
+}
+
+export interface DeleteTarget {
+  path: string;
+  name: string;
+  isDir: boolean;
+}
