+++ internal/service/ssh_svc/ssh.go
@@ -2,19 +2,15 @@ package ssh_svc
 
 import (
 	"bytes"
-	"crypto/rand"
-	"encoding/hex"
 	"fmt"
 	"io"
 	"net"
 	"os"
-	"path"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/opskat/opskat/internal/model/entity/asset_entity"
+	"github.com/opskat/opskat/internal/pkg/dirsync"
 
 	"github.com/cago-frame/cago/pkg/logger"
 	"go.uber.org/zap"
@@ -81,6 +77,7 @@ type Session struct {
 	pendingDirChange   chan error
 	pendingDirNonce    string
 	pendingDirTarget   string
+	pendingDirExpected string
 	parserRemainder    []byte
 	syncToken          string
 	promptNonce        string
@@ -90,39 +87,6 @@ type Session struct {
 	syncProbeActive    bool
 	probeShellStateFn  func(int) (shellProbeResult, error)
 }
-
-// DirectorySyncState 表示终端目录同步状态。
-type DirectorySyncState struct {
-	SessionID   string `json:"sessionId"`
-	Cwd         string `json:"cwd,omitempty"`
-	CwdKnown    bool   `json:"cwdKnown"`
-	Shell       string `json:"shell,omitempty"`
-	ShellType   string `json:"shellType,omitempty"`
-	Supported   bool   `json:"supported"`
-	PromptReady bool   `json:"promptReady"`
-	PromptClean bool   `json:"promptClean"`
-	Busy        bool   `json:"busy"`
-	Status      string `json:"status"` // "initializing" | "ready" | "unsupported"
-	LastError   string `json:"lastError,omitempty"`
-}
-
-const (
-	shellTypeUnsupported = "unsupported"
-	shellTypeBash        = "bash"
-	shellTypeZsh         = "zsh"
-	shellTypeKsh         = "ksh"
-	shellTypeMksh        = "mksh"
-
-	directorySyncInitializing = "initializing"
-	directorySyncReady        = "ready"
-	directorySyncUnsupported  = "unsupported"
-
-	syncSequencePrefix          = "\x1b]1337;opskat:"
-	syncSequenceTerm            = "\a"
-	syncSequenceParserMaxBytes  = 8 * 1024
-	syncSequenceTokenBytes      = 16
-	directorySyncMarkerOverflow = "DIRSYNC_MARKER_OVERFLOW"
-)
 
 // Write 向终端写入数据（用户输入）
 func (s *Session) Write(data []byte) error {
@@ -153,7 +117,7 @@ func (s *Session) Resize(cols, rows int) error {
 
 // Close 关闭会话
 func (s *Session) Close() {
-	s.failPendingDirectoryChange(fmt.Errorf("DIRSYNC_SESSION_CLOSED"))
+	s.failPendingDirectoryChange(dirsync.Error(dirSyncErrSessionClosed))
 	s.mu.Lock()
 	defer s.mu.Unlock()
 	if s.closed {
@@ -181,40 +145,6 @@ func (s *Session) IsClosed() bool {
 	return s.closed
 }
 
-// GetSyncState 返回目录同步状态快照。
-func (s *Session) GetSyncState() DirectorySyncState {
-	s.syncMu.Lock()
-	defer s.syncMu.Unlock()
-	return s.syncState
-}
-
-// ChangeDirectory 在终端提示符可用时切换目录，并等待 shell 确认结果。
-func (s *Session) ChangeDirectory(targetPath string) error {
-	if targetPath == "" {
-		return fmt.Errorf("DIRSYNC_INVALID_TARGET")
-	}
-
-	resultCh := make(chan error, 1)
-	command, err := s.prepareDirectoryChange(targetPath, resultCh)
-	if err != nil {
-		return err
-	}
-
-	if err := s.writeInternal([]byte(command)); err != nil {
-		s.failPendingDirectoryChange(err)
-		return err
-	}
-	s.ensureSyncProbe()
-
-	select {
-	case result := <-resultCh:
-		return result
-	case <-time.After(4 * time.Second):
-		s.failPendingDirectoryChange(fmt.Errorf("DIRSYNC_TIMEOUT"))
-		return fmt.Errorf("DIRSYNC_TIMEOUT")
-	}
-}
-
 func (s *Session) writeInternal(data []byte) error {
 	s.mu.Lock()
 	defer s.mu.Unlock()
@@ -223,338 +153,6 @@ func (s *Session) writeInternal(data []byte) error {
 	}
 	_, err := s.stdin.Write(data)
 	return err
-}
-
-func (s *Session) initSyncState(shellPath, shellType string, supported bool) {
-	state := DirectorySyncState{
-		SessionID:   s.ID,
-		Shell:       shellPath,
-		ShellType:   shellType,
-		Supported:   supported,
-		PromptReady: false,
-		PromptClean: true,
-		Status:      directorySyncUnsupported,
-	}
-	if supported {
-		state.Status = directorySyncInitializing
-	}
-	state.Busy = !state.PromptReady || !state.PromptClean
-
-	s.syncMu.Lock()
-	s.syncState = state
-	s.syncDirty = supported
-	s.syncMu.Unlock()
-	s.emitSyncState(state)
-}
-
-func (s *Session) markUserInput(data []byte) {
-	if len(data) == 0 {
-		return
-	}
-
-	s.syncMu.Lock()
-	if !s.syncState.Supported {
-		s.syncMu.Unlock()
-		return
-	}
-
-	hasNewline := bytes.ContainsAny(data, "\r\n")
-	changed := false
-	if s.syncState.PromptReady {
-		if s.syncState.PromptClean {
-			s.syncState.PromptClean = false
-			changed = true
-		}
-		if hasNewline {
-			s.syncState.PromptReady = false
-			s.syncState.CwdKnown = false
-			s.syncState.Cwd = ""
-			s.syncState.Status = directorySyncInitializing
-			s.syncDirty = true
-			changed = true
-		}
-	}
-	if changed {
-		s.syncState.Busy = !s.syncState.PromptReady || !s.syncState.PromptClean
-		state := s.syncState
-		go s.emitSyncState(state)
-	}
-	s.syncMu.Unlock()
-}
-
-func (s *Session) notePrompt(cwd string) {
-	s.syncMu.Lock()
-	s.syncState.Cwd = strings.TrimRight(cwd, "\r\n")
-	s.syncState.CwdKnown = s.syncState.Cwd != ""
-	s.syncState.PromptReady = true
-	s.syncState.PromptClean = true
-	s.syncState.Busy = false
-	s.syncState.Status = directorySyncReady
-	s.syncState.LastError = ""
-	s.syncDirty = false
-	state := s.syncState
-	s.syncMu.Unlock()
-	s.emitSyncState(state)
-}
-
-func (s *Session) noteObservedCwd(cwd string) {
-	cleaned := strings.TrimRight(cwd, "\r\n")
-	if cleaned == "" {
-		return
-	}
-
-	s.syncMu.Lock()
-	s.syncState.Cwd = cleaned
-	s.syncState.CwdKnown = true
-	s.syncDirty = false
-	state := s.syncState
-	s.syncMu.Unlock()
-	s.emitSyncState(state)
-}
-
-func (s *Session) prepareDirectoryChange(targetPath string, resultCh chan error) (string, error) {
-	s.syncMu.Lock()
-	defer s.syncMu.Unlock()
-
-	switch {
-	case !s.syncState.Supported:
-		return "", fmt.Errorf("DIRSYNC_UNSUPPORTED")
-	case !s.syncState.CwdKnown:
-		return "", fmt.Errorf("DIRSYNC_CWD_UNKNOWN")
-	case s.pendingDirChange != nil:
-		return "", fmt.Errorf("DIRSYNC_PENDING")
-	case !s.syncState.PromptReady || !s.syncState.PromptClean:
-		return "", fmt.Errorf("DIRSYNC_BUSY")
-	}
-
-	nonce, err := generateSyncToken()
-	if err != nil {
-		return "", fmt.Errorf("DIRSYNC_NONCE_FAILED")
-	}
-	s.pendingDirChange = resultCh
-	s.pendingDirNonce = nonce
-	s.pendingDirTarget = targetPath
-	s.syncState.PromptReady = false
-	s.syncState.PromptClean = false
-	s.syncState.CwdKnown = false
-	s.syncState.Cwd = ""
-	s.syncState.Busy = true
-	s.syncState.Status = directorySyncInitializing
-	s.syncState.LastError = ""
-	s.syncDirty = true
-	state := s.syncState
-
-	go s.emitSyncState(state)
-	return buildDirectoryChangeCommand(targetPath), nil
-}
-
-func (s *Session) finishDirectoryChange(err error, cwd string) {
-	s.syncMu.Lock()
-	ch := s.pendingDirChange
-	s.pendingDirChange = nil
-	s.pendingDirNonce = ""
-	s.pendingDirTarget = ""
-	if cwd != "" {
-		s.syncState.Cwd = strings.TrimRight(cwd, "\r\n")
-		s.syncState.CwdKnown = s.syncState.Cwd != ""
-		s.syncState.PromptReady = true
-		s.syncState.PromptClean = true
-		s.syncState.Busy = false
-		s.syncState.Status = directorySyncReady
-	}
-	if err != nil {
-		s.syncState.LastError = err.Error()
-	} else {
-		s.syncState.LastError = ""
-	}
-	s.syncDirty = false
-	state := s.syncState
-	s.syncMu.Unlock()
-
-	if ch != nil {
-		ch <- err
-		close(ch)
-	}
-	s.emitSyncState(state)
-}
-
-func (s *Session) failPendingDirectoryChange(err error) {
-	s.syncMu.Lock()
-	ch := s.pendingDirChange
-	s.pendingDirChange = nil
-	s.pendingDirNonce = ""
-	s.pendingDirTarget = ""
-	if err != nil {
-		s.syncState.LastError = err.Error()
-	}
-	state := s.syncState
-	s.syncMu.Unlock()
-
-	if ch != nil {
-		ch <- err
-		close(ch)
-	}
-	s.emitSyncState(state)
-}
-
-func (s *Session) emitSyncState(state DirectorySyncState) {
-	if s.onSync == nil {
-		return
-	}
-	s.onSync(s.ID, state)
-}
-
-func (s *Session) noteParserOverflow() {
-	s.syncMu.Lock()
-	if s.syncState.LastError == directorySyncMarkerOverflow {
-		s.syncMu.Unlock()
-		return
-	}
-	s.syncState.LastError = directorySyncMarkerOverflow
-	state := s.syncState
-	s.syncMu.Unlock()
-	s.emitSyncState(state)
-}
-
-type shellProbeResult struct {
-	cwd         string
-	promptReady bool
-}
-
-func (s *Session) ensureSyncProbe() {
-	s.syncMu.Lock()
-	if s.syncProbeActive || !s.syncState.Supported || s.shellPID <= 0 || s.shared == nil || s.shared.client == nil {
-		s.syncMu.Unlock()
-		return
-	}
-	s.syncProbeActive = true
-	s.syncMu.Unlock()
-
-	go s.runSyncProbeLoop()
-}
-
-func (s *Session) runSyncProbeLoop() {
-	ticker := time.NewTicker(250 * time.Millisecond)
-	defer ticker.Stop()
-
-	for {
-		s.syncMu.Lock()
-		if s.closed || !s.syncState.Supported || s.shellPID <= 0 || s.shared == nil || s.shared.client == nil {
-			s.syncProbeActive = false
-			s.syncMu.Unlock()
-			return
-		}
-		shouldProbe := s.syncDirty || s.pendingDirChange != nil
-		pid := s.shellPID
-		pending := s.pendingDirChange != nil
-		pendingNonce := s.pendingDirNonce
-		pendingTarget := s.pendingDirTarget
-		s.syncMu.Unlock()
-
-		if !shouldProbe {
-			s.syncMu.Lock()
-			s.syncProbeActive = false
-			s.syncMu.Unlock()
-			return
-		}
-
-		result, err := s.probeShellState(pid)
-		if err == nil {
-			if pending {
-				s.finishPendingDirectoryChangeProbe(pendingNonce, pendingTarget, result.cwd)
-			} else if result.cwd != "" {
-				s.noteObservedCwd(result.cwd)
-			}
-		}
-
-		<-ticker.C
-	}
-}
-
-func (s *Session) finishPendingDirectoryChangeProbe(nonce, targetPath, cwd string) {
-	s.syncMu.Lock()
-	if s.pendingDirChange == nil || s.pendingDirNonce == "" || s.pendingDirNonce != nonce {
-		s.syncMu.Unlock()
-		return
-	}
-	s.syncMu.Unlock()
-
-	if cwd == "" {
-		return
-	}
-	if path.Clean(cwd) == path.Clean(targetPath) {
-		s.finishDirectoryChange(nil, cwd)
-	}
-}
-
-func (s *Session) probeShellState(shellPID int) (shellProbeResult, error) {
-	if s.probeShellStateFn != nil {
-		return s.probeShellStateFn(shellPID)
-	}
-	session, err := s.shared.client.NewSession()
-	if err != nil {
-		return shellProbeResult{}, err
-	}
-	defer func() {
-		if closeErr := session.Close(); closeErr != nil && closeErr != io.EOF {
-			logger.Default().Warn("close shell probe session", zap.Error(closeErr))
-		}
-	}()
-
-	var out bytes.Buffer
-	session.Stdout = &out
-	session.Stderr = io.Discard
-	if err := session.Run(buildShellStateProbeCommand(shellPID)); err != nil {
-		return shellProbeResult{}, err
-	}
-	return parseShellProbeOutput(out.Bytes())
-}
-
-func buildShellStateProbeCommand(shellPID int) string {
-	return fmt.Sprintf(`sh -lc 'pid=%d
-cwd=""
-prompt=0
-if kill -0 "$pid" 2>/dev/null; then
-  cwd=$(readlink "/proc/$pid/cwd" 2>/dev/null || printf "")
-  if [ -z "$cwd" ] && command -v pwdx >/dev/null 2>&1; then
-    cwd=$(pwdx "$pid" 2>/dev/null | sed "s/^[^ ]* //")
-  fi
-  pgid=$(ps -o pgid= -p "$pid" 2>/dev/null | tr -d " ")
-  tpgid=$(ps -o tpgid= -p "$pid" 2>/dev/null | tr -d " ")
-  tty_path=$(readlink "/proc/$pid/fd/0" 2>/dev/null || printf "")
-  if [ -n "$tty_path" ]; then
-    stty_state=$(stty -a < "$tty_path" 2>/dev/null || printf "")
-    case "$stty_state" in
-      *"-icanon"*"-echo"*)
-        if [ -n "$pgid" ] && [ "$pgid" = "$tpgid" ]; then
-          prompt=1
-        fi
-        ;;
-    esac
-  fi
-fi
-printf "cwd=%%s\0prompt=%%s\0" "$cwd" "$prompt"'`, shellPID)
-}
-
-func parseShellProbeOutput(raw []byte) (shellProbeResult, error) {
-	result := shellProbeResult{}
-	fields := bytes.Split(raw, []byte{0})
-	for _, field := range fields {
-		if len(field) == 0 {
-			continue
-		}
-		key, value, ok := bytes.Cut(field, []byte{'='})
-		if !ok {
-			return shellProbeResult{}, fmt.Errorf("invalid probe field")
-		}
-		switch string(key) {
-		case "cwd":
-			result.cwd = string(value)
-		case "prompt":
-			result.promptReady = string(value) == "1"
-		}
-	}
-	return result, nil
 }
 
 // Manager 管理所有 SSH 会话
@@ -797,150 +395,6 @@ func (m *Manager) NewSessionFrom(existingSessionID string, cols, rows int,
 	}
 
 	return sessionID, nil
-}
-
-func detectRemoteShell(client *ssh.Client) (string, string) {
-	session, err := client.NewSession()
-	if err != nil {
-		return "/bin/sh", shellTypeUnsupported
-	}
-	defer func() {
-		if closeErr := session.Close(); closeErr != nil && closeErr != io.EOF {
-			logger.Default().Warn("close shell probe session", zap.Error(closeErr))
-		}
-	}()
-
-	var out bytes.Buffer
-	session.Stdout = &out
-	session.Stderr = io.Discard
-	if err := session.Run(`sh -lc 'printf "%s" "${SHELL:-/bin/sh}"'`); err != nil {
-		return "/bin/sh", shellTypeUnsupported
-	}
-
-	shellPath := strings.TrimSpace(out.String())
-	if shellPath == "" {
-		shellPath = "/bin/sh"
-	}
-	return shellPath, normalizeShellType(shellPath)
-}
-
-func normalizeShellType(shellPath string) string {
-	switch path.Base(shellPath) {
-	case "bash":
-		return shellTypeBash
-	case "zsh":
-		return shellTypeZsh
-	case "ksh":
-		return shellTypeKsh
-	case "mksh":
-		return shellTypeMksh
-	default:
-		return shellTypeUnsupported
-	}
-}
-
-func generateSyncToken() (string, error) {
-	buf := make([]byte, syncSequenceTokenBytes)
-	if _, err := rand.Read(buf); err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(buf), nil
-}
-
-func buildInteractiveShellCommand(shellPath, shellType, syncToken, promptNonce string) string {
-	switch shellType {
-	case shellTypeBash:
-		return fmt.Sprintf(`rc="$(mktemp "${TMPDIR:-/tmp}/opskat-bash-XXXXXX")" && cat >"$rc" <<'EOF'
-if [ -f "$HOME/.bash_profile" ]; then
-  . "$HOME/.bash_profile"
-elif [ -f "$HOME/.bash_login" ]; then
-  . "$HOME/.bash_login"
-elif [ -f "$HOME/.profile" ]; then
-  . "$HOME/.profile"
-fi
-[ -f "$HOME/.bashrc" ] && . "$HOME/.bashrc"
-opskat_next_prompt_nonce() {
-  local opskat_now opskat_rand
-  opskat_now=$(date +%%s%%N 2>/dev/null || date +%%s 2>/dev/null || printf '0')
-  opskat_rand=${RANDOM:-0}
-  printf '%%s-%%s-%%s' "$$" "$opskat_rand" "$opskat_now"
-}
-opskat_prompt_proof() {
-  local opskat_pwd opskat_current opskat_next
-  opskat_current=${OPSKAT_PROMPT_NONCE:-}
-  [ -n "$opskat_current" ] || return
-  opskat_next=$(opskat_next_prompt_nonce)
-  opskat_pwd=$(builtin pwd -P 2>/dev/null || builtin pwd 2>/dev/null || printf '')
-  printf '\033]1337;opskat:%s:prompt:%%s:%%s:%%s\007' "$opskat_current" "$opskat_next" "$opskat_pwd"
-  OPSKAT_PROMPT_NONCE=$opskat_next
-}
-OPSKAT_PROMPT_NONCE=%s
-PROMPT_COMMAND="opskat_prompt_proof${PROMPT_COMMAND:+;$PROMPT_COMMAND}"
-EOF
-printf '\033]1337;opskat:%s:init:pid:%%s\007' "$$"
-exec %s --rcfile "$rc" -i`, syncToken, shellQuote(promptNonce), syncToken, shellQuote(shellPath))
-	case shellTypeZsh:
-		return fmt.Sprintf(`dir="$(mktemp -d "${TMPDIR:-/tmp}/opskat-zsh-XXXXXX")" && cat >"$dir/.zshenv" <<'EOF_ENV'
-[[ -f "$HOME/.zshenv" ]] && source "$HOME/.zshenv"
-EOF_ENV
-cat >"$dir/.zshrc" <<'EOF_RC'
-[[ -f "$HOME/.zprofile" ]] && source "$HOME/.zprofile"
-[[ -f "$HOME/.zshrc" ]] && source "$HOME/.zshrc"
-autoload -Uz add-zsh-hook
-opskat_next_prompt_nonce() {
-  local opskat_now opskat_rand
-  opskat_now=$(date +%%s%%N 2>/dev/null || date +%%s 2>/dev/null || printf '0')
-  opskat_rand=${RANDOM:-0}
-  printf '%%s-%%s-%%s' "$$" "$opskat_rand" "$opskat_now"
-}
-opskat_prompt_proof() {
-  local opskat_pwd opskat_current opskat_next
-  opskat_current=${OPSKAT_PROMPT_NONCE:-}
-  [[ -n "$opskat_current" ]] || return
-  opskat_next=$(opskat_next_prompt_nonce)
-  opskat_pwd=$(pwd -P 2>/dev/null || pwd 2>/dev/null || printf '')
-  printf '\033]1337;opskat:%s:prompt:%%s:%%s:%%s\007' "$opskat_current" "$opskat_next" "$opskat_pwd"
-  OPSKAT_PROMPT_NONCE=$opskat_next
-}
-OPSKAT_PROMPT_NONCE=%s
-add-zsh-hook precmd opskat_prompt_proof
-EOF_RC
-export ZDOTDIR="$dir"
-printf '\033]1337;opskat:%s:init:pid:%%s\007' "$$"
-exec %s -i`, syncToken, shellQuote(promptNonce), syncToken, shellQuote(shellPath))
-	case shellTypeKsh, shellTypeMksh:
-		return fmt.Sprintf(`envfile="$(mktemp "${TMPDIR:-/tmp}/opskat-ksh-XXXXXX")" && cat >"$envfile" <<'EOF'
-[ -f "$HOME/.profile" ] && . "$HOME/.profile"
-opskat_next_prompt_nonce() {
-  OPSKAT_NOW=$(date +%%s%%N 2>/dev/null || date +%%s 2>/dev/null || printf '0')
-  OPSKAT_RAND=${RANDOM:-0}
-  printf '%%s-%%s-%%s' "$$" "$OPSKAT_RAND" "$OPSKAT_NOW"
-}
-opskat_prompt_proof() {
-  OPSKAT_CURRENT=${OPSKAT_PROMPT_NONCE:-}
-  [ -n "$OPSKAT_CURRENT" ] || return
-  OPSKAT_NEXT=$(opskat_next_prompt_nonce)
-  OPSKAT_PWD=$(pwd -P 2>/dev/null || pwd 2>/dev/null || printf '')
-  printf '\033]1337;opskat:%s:prompt:%%s:%%s:%%s\007' "$OPSKAT_CURRENT" "$OPSKAT_NEXT" "$OPSKAT_PWD"
-  OPSKAT_PROMPT_NONCE=$OPSKAT_NEXT
-}
-OPSKAT_PROMPT_NONCE=%s
-PS1='$(opskat_prompt_proof)'"$PS1"
-EOF
-export ENV="$envfile"
-printf '\033]1337;opskat:%s:init:pid:%%s\007' "$$"
-exec %s -i`, syncToken, shellQuote(promptNonce), syncToken, shellQuote(shellPath))
-	default:
-		return ""
-	}
-}
-
-func buildDirectoryChangeCommand(targetPath string) string {
-	return fmt.Sprintf("builtin cd -- %s\r", shellQuote(targetPath))
-}
-
-func shellQuote(value string) string {
-	return "'" + strings.ReplaceAll(value, "'", `'\''`) + "'"
 }
 
 // dial 建立到目标的网络连接，支持代理和跳板机链
@@ -1287,141 +741,6 @@ func (m *Manager) readOutput(sess *Session) {
 	}
 }
 
-func (s *Session) filterOutput(chunk []byte) []byte {
-	data := chunk
-	if len(s.parserRemainder) > 0 {
-		data = append(append([]byte(nil), s.parserRemainder...), chunk...)
-		s.parserRemainder = nil
-	}
-
-	prefix := []byte(syncSequencePrefix)
-	out := make([]byte, 0, len(data))
-
-	for len(data) > 0 {
-		idx := bytes.Index(data, prefix)
-		if idx < 0 {
-			break
-		}
-		out = append(out, data[:idx]...)
-		remainder := data[idx+len(prefix):]
-		end := bytes.IndexByte(remainder, syncSequenceTerm[0])
-		if end < 0 {
-			tail := append([]byte(nil), data[idx:]...)
-			if len(tail) > syncSequenceParserMaxBytes {
-				s.noteParserOverflow()
-				out = append(out, tail...)
-				return out
-			}
-			s.parserRemainder = tail
-			return out
-		}
-		rawEnd := idx + len(prefix) + end + 1
-		raw := data[idx:rawEnd]
-		if !s.handleSyncPayload(string(remainder[:end])) {
-			out = append(out, raw...)
-		}
-		data = data[rawEnd:]
-	}
-
-	if len(data) == 0 {
-		return out
-	}
-
-	if keep := trailingPrefixLength(data, prefix); keep > 0 {
-		out = append(out, data[:len(data)-keep]...)
-		s.parserRemainder = append([]byte(nil), data[len(data)-keep:]...)
-		return out
-	}
-
-	out = append(out, data...)
-	return out
-}
-
-func trailingPrefixLength(data, prefix []byte) int {
-	maxSize := len(prefix) - 1
-	if maxSize > len(data) {
-		maxSize = len(data)
-	}
-	for size := maxSize; size > 0; size-- {
-		if bytes.Equal(data[len(data)-size:], prefix[:size]) {
-			return size
-		}
-	}
-	return 0
-}
-
-func (s *Session) handleSyncPayload(payload string) bool {
-	token, body, ok := strings.Cut(payload, ":")
-	if !ok || token == "" || token != s.syncToken {
-		return false
-	}
-
-	switch {
-	case strings.HasPrefix(body, "init:pid:"):
-		pidText := strings.TrimPrefix(body, "init:pid:")
-		pid, err := strconv.Atoi(strings.TrimSpace(pidText))
-		if err != nil || pid <= 0 {
-			return false
-		}
-		s.syncMu.Lock()
-		if s.shellPID != 0 {
-			s.syncMu.Unlock()
-			return false
-		}
-		s.shellPID = pid
-		s.syncDirty = true
-		s.syncMu.Unlock()
-		s.ensureSyncProbe()
-		return true
-	case strings.HasPrefix(body, "prompt:"):
-		remainder := strings.TrimPrefix(body, "prompt:")
-		currentNonce, nextPayload, ok := strings.Cut(remainder, ":")
-		if !ok || currentNonce == "" {
-			return false
-		}
-		nextNonce, cwd, ok := strings.Cut(nextPayload, ":")
-		if !ok || nextNonce == "" {
-			return false
-		}
-		s.syncMu.Lock()
-		promptNonce := s.promptNonce
-		promptPendingNonce := s.promptPendingNonce
-		shellPID := s.shellPID
-		s.syncMu.Unlock()
-		validCurrent := currentNonce == promptNonce || (promptPendingNonce != "" && currentNonce == promptPendingNonce)
-		if promptNonce == "" || !validCurrent || shellPID <= 0 {
-			return false
-		}
-		probe, err := s.probeShellState(shellPID)
-		if err != nil || !probe.promptReady {
-			s.syncMu.Lock()
-			if currentNonce == s.promptNonce || (s.promptPendingNonce != "" && currentNonce == s.promptPendingNonce) {
-				s.promptPendingNonce = nextNonce
-			}
-			s.syncMu.Unlock()
-			return false
-		}
-		resolvedCwd := probe.cwd
-		if resolvedCwd == "" {
-			resolvedCwd = cwd
-		}
-		if resolvedCwd == "" {
-			return false
-		}
-		s.syncMu.Lock()
-		if currentNonce != s.promptNonce && (s.promptPendingNonce == "" || currentNonce != s.promptPendingNonce) {
-			s.syncMu.Unlock()
-			return false
-		}
-		s.promptNonce = nextNonce
-		s.promptPendingNonce = ""
-		s.syncMu.Unlock()
-		s.notePrompt(resolvedCwd)
-		return true
-	}
-	return false
-}
-
 // GetSession 获取会话
 func (m *Manager) GetSession(id string) (*Session, bool) {
 	v, ok := m.sessions.Load(id)
@@ -1435,7 +754,7 @@ func (m *Manager) GetSession(id string) (*Session, bool) {
 func (m *Manager) GetSessionSyncState(id string) (DirectorySyncState, error) {
 	sess, ok := m.GetSession(id)
 	if !ok {
-		return DirectorySyncState{}, fmt.Errorf("会话不存在: %s", id)
+		return DirectorySyncState{}, dirsync.Error(dirsync.CodeSessionNotFound)
 	}
 	return sess.GetSyncState(), nil
 }
